// src/app/models/ortodonciaseguimiento.interface.ts
export interface SeguimientoOrtodonciaInterface {
  id?: string;
  idOrtodoncia?: string;
  ComentariodegOGNPMAX?: string;
  ComentariodesUPINF?: string;
  ComentariodeaNB?: string;
  ComentariodesNB?: string;
  ComentariodesNA?: string;
  ComentariodesNGN?: string;
  ComentariodenAPOG?: string;
  ComentariodesNGOGN?: string;
  ComentariodesNPLO?: string;
  ComentariodesNFH?: string;
  Fecha?: string;
  TratamientoEfectuado?: string;
  Observaciones?: string;
  ConsultasPosteriores?: string;
  Estatus?: string;


}